+++ CHANGES.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+development version
+-------------------
+
+* Added support for chunking FASTA reads to ``read_paired_chunks``.
+  Previously, only FASTQ was supported.
+
 v1.1.0 (2023-11-20)
 -------------------